+++ developerRobot/src/main/java/frc/robot/Robot.java
@@ -4,6 +4,7 @@
 
 package frc.robot;
 
+import edu.wpi.first.wpilibj.Compressor;
 import edu.wpi.first.wpilibj.PneumaticsModuleType;
 import edu.wpi.first.wpilibj.Solenoid;
 import edu.wpi.first.wpilibj.TimedRobot;
@@ -16,24 +17,24 @@ public class Robot extends TimedRobot {
   @Override
   @SuppressWarnings("resource")
   public void robotInit() {
-    Solenoid sCTRE1 = new Solenoid(PneumaticsModuleType.CTREPCM, 0);
-    Solenoid sREV1 = new Solenoid(PneumaticsModuleType.REVPH, 0);
-    Solenoid sREV2 = new Solenoid(PneumaticsModuleType.REVPH, 12);
-    sCTRE1.set(true);
-    sREV1.set(true);
-    sREV2.set(true);
+    // Solenoid sCTRE1 = new Solenoid(PneumaticsModuleType.CTREPCM, 0);
+    // Solenoid sREV1 = new Solenoid(PneumaticsModuleType.REVPH, 0);
+    // Solenoid sREV2 = new Solenoid(PneumaticsModuleType.REVPH, 12);
+    // sCTRE1.set(true);
+    // sREV1.set(true);
+    // sREV2.set(true);
     // new Compressor(PneumaticsModuleType.CTREPCM);/
-    // new Compressor(PneumaticsModuleType.REVPH);
+    new Compressor(PneumaticsModuleType.REVPH);
   }
 
   /*
    * Good
    * main.cpp
    * PHSimGui.h
-   * Getting there, still need to split compressor data, consolidate solenoid views?, rename PCM.cpp and PCM.h
+   * Getting there, still need to split compressor data, consolidate solenoid views?
    * PHSimGui.cpp
-   * PCM.cpp
-   * PCM.h
+   * Pneumatic.cpp
+   * Pneumatic.h
    * 
    */
