+++ glass/src/lib/native/cpp/hardware/Pneumatic.cpp
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "glass/hardware/PCM.h"
+#include "glass/hardware/Pneumatic.h"
 
 #include <cstdio>
 #include <cstring>
