+++ simulation/halsim_gui/src/main/native/cpp/PCMSimGui.cpp
@@ -4,7 +4,7 @@
 
 #include "PCMSimGui.h"
 
-#include <glass/hardware/PCM.h>
+#include <glass/hardware/Pneumatic.h>
 #include <glass/other/DeviceTree.h>
 
 #include <cstdio>
